Priorities now shorts to fix int overflow when comparing

+++ IconsAddon/src/main/java/IconsAddon/blockModifiers/AbstractBlockModifier.java
@@ -144,7 +144,7 @@ public abstract class AbstractBlockModifier implements Comparable<AbstractBlockM
         return Priority.NORMAL;
     }
 
-    public int subPriority() {
+    public short subPriority() {
         return 0;
     }
